Refactor some common AllocatorTestBase paramaters that will be used across multiple classes as part of the integration test updates.
